+++ backend/src/Logitar.Portal.Application/Sessions/Commands/SignInSessionCommandHandler.cs
@@ -3,10 +3,9 @@ using Logitar.EventSourcing;
 using Logitar.Identity.Contracts.Settings;
 using Logitar.Identity.Domain.Passwords;
 using Logitar.Identity.Domain.Sessions;
-using Logitar.Identity.Domain.Shared;
 using Logitar.Identity.Domain.Users;
 using Logitar.Portal.Application.Sessions.Validators;
-using Logitar.Portal.Application.Users;
+using Logitar.Portal.Application.Users.Queries;
 using Logitar.Portal.Contracts;
 using Logitar.Portal.Contracts.Sessions;
 using MediatR;
@@ -15,14 +14,16 @@ namespace Logitar.Portal.Application.Sessions.Commands;
 
 internal class SignInSessionCommandHandler : IRequestHandler<SignInSessionCommand, Session>
 {
+  private readonly IMediator _mediator;
   private readonly IPasswordManager _passwordManager;
   private readonly ISessionQuerier _sessionQuerier;
   private readonly ISessionRepository _sessionRepository;
   private readonly IUserManager _userManager;
 
-  public SignInSessionCommandHandler(IPasswordManager passwordManager,
+  public SignInSessionCommandHandler(IMediator mediator, IPasswordManager passwordManager,
     ISessionQuerier sessionQuerier, ISessionRepository sessionRepository, IUserManager userManager)
   {
+    _mediator = mediator;
     _passwordManager = passwordManager;
     _sessionQuerier = sessionQuerier;
     _sessionRepository = sessionRepository;
@@ -36,9 +37,8 @@ internal class SignInSessionCommandHandler : IRequestHandler<SignInSessionComman
     SignInSessionPayload payload = command.Payload;
     new SignInSessionValidator().ValidateAndThrow(payload);
 
-    TenantId? tenantId = command.TenantId;
-    FoundUsers users = await _userManager.FindAsync(tenantId?.Value, payload.UniqueName, userSettings, cancellationToken);
-    UserAggregate user = users.SingleOrDefault() ?? throw new UserNotFoundException(tenantId, payload.UniqueName, nameof(payload.UniqueName));
+    FindUserQuery query = new(command.TenantId, payload.UniqueName, userSettings, nameof(payload.UniqueName));
+    UserAggregate user = await _mediator.Send(query, cancellationToken);
     ActorId actorId = new(user.Id.Value);
 
     Password? secret = null;